+++ mysql2ch/reader.py
@@ -5,12 +5,12 @@ from copy import deepcopy
 
 import MySQLdb
 from MySQLdb.cursors import DictCursor
-
-from mysql2ch.common import complex_decode
-from mysql2ch.convert import SqlConvert
 from pymysqlreplication import BinLogStreamReader
 from pymysqlreplication.event import QueryEvent
 from pymysqlreplication.row_event import DeleteRowsEvent, UpdateRowsEvent, WriteRowsEvent
+
+from mysql2ch.common import complex_decode
+from mysql2ch.convert import SqlConvert
 
 logger = logging.getLogger("mysql2ch.reader")